Publish generic Omnigent repository work
